Implementation of task creation optimization
The application level uses system calls to add tasks to the kernel's task queue,
ensuring that scheduling of tasks is done only at the kernel level

+++ screen.c
@@ -24,7 +24,7 @@ void ClearScreen()
     SetPrintPos(0, 0);
 }
 
-int  SetPrintPos(short w, short h)
+int  SetPrintPos(byte w, byte h)
 {
     int ret = 0;
