+++ screen.h
@@ -15,11 +15,11 @@ typedef enum
 } PrintColor;
 
 void ClearScreen();
-int  SetPrintPos(short w, short h);
+int  SetPrintPos(byte w, byte h);
 void SetPrintColor(PrintColor c);
 int PrintChar(char c);
 int PrintString(const char* s);
 int PrintIntDec(int n);
-int PrintIntHex(unsigned int n);
+int PrintIntHex(uint n);
 
 #endif
